refactor: change mock fetch to be more specific
