+++ package.json
@@ -38,7 +38,7 @@
       "src/**/*.{js,jsx}",
       "!src/index.js",
       "!src/serviceWorker.js",
-      "!src/unit-test/mocks/fetch.js"
+      "!src/unit-test/mocks/recipeListFetch.js"
     ]
   }
 }
