+++ src/feature-test/user_sees_expanded_recipe_when_clicked.test.js
@@ -1,5 +1,5 @@
 import '../setupTest.js'
-import fetch from '../unit-test/mocks/fetch';
+import fetch from '../unit-test/mocks/recipeListFetch';
 
 describe('I.I.A.S', () => {
   it('the user sees expanded recipe when clicking on title', async () => {
